Prime Number Generator Code Challenge
Build a simple program that generates primes from given start and end inputs using sympy. Built a couple tests using pytest. Instructions on how to run the Prime Number Generator and the tests are in the README.md file as well as my thought process on building and testing the program.